consolidate some test utilities in utils
instead of in test_api, test_pages

since they are used in a few places

also add user, username fixtures for generating test users
